board

+++ jsp-lecture/jsp04/src/main/webapp/board/view.jsp
@@ -56,22 +56,22 @@ ResultSet rs = pstmt.executeQuery();
 						<th>내용</th>
 						<td colspan="3" class="content"><%=rs.getString("content")%></td>
 					</tr>
-
 				</tbody>
-				
 			</table>
 			<div class="d-flex justify-content-center mt-5">
 				<a href="../board/list.jsp" class="btn btn-primary">목록</a>
 				<a href="../board/write.jsp" class="btn btn-primary mx-1">글쓰기</a>
-				<a href="../board/delete.jsp?no=<%=rs.getInt("no") %>" class="btn btn-danger mx-1">지우기</a>
+				<%if(loggedID.equals(rs.getString("id"))){ %>
+					<a href="../board/delete.jsp?no=<%=rs.getInt("no") %>" class="btn btn-danger mx-1">지우기</a>
+					<a href="../board/modify.jsp?no=<%=rs.getInt("no") %>" class="btn btn-danger mx-1">수정하기</a>
+				<%} %>
 			</div>
-			<%
-				}
-			%>
+			
 		</div>
 	</div>
 </div>
 <%
+}
 jdbcConn.close();
 %>
 <%@ include file="../include/footer.jsp"%>
